Change signature of update methods

+++ src/main/java/it/unibo/pyxis/model/arena/Arena.java
@@ -17,7 +17,7 @@ public interface Arena {
      * @param delta
      *              The passed time.
      */
-    void update(int delta);
+    void update(double delta);
 
     /**
      * Handle a {@link BrickDestructionEvent}.
